Refactor mongoDB seed script to findOneAndUpdate, reuse code
